Move assert_bad_user_input function to test.utils module
